fix(flink): Make service monitor names unique per cluster
Helm 3 will not install the chart a second time due to the name conflict

"Unable to continue with install: ServiceMonitor "flink-jobmanager" in
namespace "default" exists and cannot be imported into the current
release"

Helm 2 will install the second cluster but only one set of service
monitors will exist (I haven't checked if the first cluster's monitors
are overwritten or if the second's are not created). When the second
cluster is uninstalled the service monitors are removed, breaking
monitoring for the first cluster.

Using the full name stops this from happening.