feat(flink): switch to bitnami/zookeeper chart
The incubator/zookeeper chart is deprecated and has not been updated for
Kubernetes 1.16.

This commit switches the dependency to the bitnami/zookeeper chart
which appears to be the best Zookeeper chart on Helm hub.

The README has been updated as some configuration had to be changed.
Some documentation for Zookeeper that appears out of date has also been
changed.